arreglar dependencias de maven sincronizandolos en el pom con click derecho y sinc project

+++ src/test/java/com/globant/automation/test/PetTests.java
@@ -45,7 +45,7 @@ public class PetTests extends TestRunner {
                 .get("/pet/{petId}")
                 .then()
                 .statusCode(200)
-                .body("id", equalTo(petId))
+                .body("id", equalTo((int) petId))
                 .body("name", equalTo("FirulaisPerfDog"));
     }
 }
